feat: implemented scroll-view enhancer
