+++ packages/uikit-chat-hooks/src/__tests__/channel/useGroupChannelMessages/useGroupChannelMessagesWithQuery.test.ts
@@ -18,7 +18,7 @@ describe('useGroupChannelMessagesWithQuery', () => {
       expect(result.current.loading).toBe(false);
       expect(result.current.refreshing).toBe(false);
       expect(result.current.messages).not.toEqual([]);
-      expect(result.current.nextMessages).toEqual([]);
+      expect(result.current.newMessages).toEqual([]);
       expect(result.current.newMessagesFromMembers).toEqual([]);
       expect(result.current.prev).toBeInstanceOf(Function);
       expect(result.current.next).toBeInstanceOf(Function);
