+++ packages/uikit-react-native/src/constants.ts
@@ -1,6 +1,6 @@
 export const DEFAULT_LONG_PRESS_DELAY = 350;
 export const DEPRECATION_WARNING = {
-  GROUP_CHANNEL: {
+  CHANNEL: {
     ON_PRESS_IMAGE_MESSAGE: '`onPressImageMessage` is deprecated, please use `onPressMediaMessage` instead',
   },
 } as const;
