+++ packages/uikit-react-native/src/localization/createBaseStringSet.ts
@@ -32,13 +32,15 @@ export const createBaseStringSet = ({ dateLocale, overrides }: StringSetCreateOp
     OPEN_CHANNEL: {
       HEADER_TITLE: (channel) => getOpenChannelTitle(channel),
       HEADER_SUBTITLE: (channel) => getOpenChannelParticipants(channel),
-      LIST_BANNER_FROZEN: 'Channel is frozen',
       LIST_DATE_SEPARATOR: (date, locale) => getDateSeparatorFormat(date, locale ?? dateLocale),
       MESSAGE_BUBBLE_TIME: (message, locale) => getMessageTimeFormat(new Date(message.createdAt), locale ?? dateLocale),
       MESSAGE_BUBBLE_FILE_TITLE: (message) => message.name,
       MESSAGE_BUBBLE_EDITED_POSTFIX: ' (edited)',
       MESSAGE_BUBBLE_UNKNOWN_TITLE: () => '(Unknown message type)',
       MESSAGE_BUBBLE_UNKNOWN_DESC: () => 'Cannot read this message.',
+
+      /** @deprecated Please use LABELS.CHANNEL_MESSAGE_LIST_FROZEN **/
+      LIST_BANNER_FROZEN: 'Channel is frozen',
       ...overrides?.OPEN_CHANNEL,
     },
     OPEN_CHANNEL_PARTICIPANTS: {
@@ -110,7 +112,6 @@ export const createBaseStringSet = ({ dateLocale, overrides }: StringSetCreateOp
     },
     GROUP_CHANNEL: {
       HEADER_TITLE: (uid, channel) => getGroupChannelTitle(uid, channel, USER_NO_NAME, CHANNEL_NO_MEMBERS),
-      LIST_BANNER_FROZEN: 'Channel is frozen',
       LIST_DATE_SEPARATOR: (date, locale) => getDateSeparatorFormat(date, locale ?? dateLocale),
       LIST_BUTTON_NEW_MSG: (newMessages) => `${newMessages.length} new messages`,
 
@@ -122,6 +123,8 @@ export const createBaseStringSet = ({ dateLocale, overrides }: StringSetCreateOp
 
       MENTION_LIMITED: (mentionLimit) => `You can have up to ${mentionLimit} mentions per message.`,
 
+      /** @deprecated Please use LABELS.CHANNEL_MESSAGE_LIST_FROZEN **/
+      LIST_BANNER_FROZEN: 'Channel is frozen',
       /** @deprecated Please use LABELS.CHANNEL_MESSAGE_COPY **/
       DIALOG_MESSAGE_COPY: 'Copy',
       /** @deprecated Please use LABELS.CHANNEL_MESSAGE_EDIT **/
@@ -286,6 +289,10 @@ export const createBaseStringSet = ({ dateLocale, overrides }: StringSetCreateOp
       UNBAN: 'Unban',
 
       // Deprecation backward
+      CHANNEL_MESSAGE_LIST_FROZEN:
+        overrides?.GROUP_CHANNEL?.LIST_BANNER_FROZEN ??
+        overrides?.OPEN_CHANNEL?.LIST_BANNER_FROZEN ??
+        'Channel is frozen',
       CHANNEL_MESSAGE_COPY: overrides?.GROUP_CHANNEL?.DIALOG_MESSAGE_COPY ?? 'Copy',
       CHANNEL_MESSAGE_EDIT: overrides?.GROUP_CHANNEL?.DIALOG_MESSAGE_EDIT ?? 'Edit',
       CHANNEL_MESSAGE_SAVE: overrides?.GROUP_CHANNEL?.DIALOG_MESSAGE_SAVE ?? 'Save',
